seperate app.py file for models and code is working
