+++ frontend/src/components/diseasePrediction/Symptom.jsx
@@ -226,7 +226,7 @@ const Symptom = forwardRef((props, ref) => {
   const sendSymptomsToBackend = () => {
     console.log("Method called!")
 
-    fetch("http://127.0.0.1:5000/predict", {
+    fetch("http://127.0.0.1:5001/predict", {
       method: "POST",
       headers: {
         "Content-Type": "application/json",
