+++ models/requirements.txt
@@ -1,0 +1,6 @@
+numpy
+pandas
+gunicorn
+scikit-learn==1.5.2
+flask-cors
+flask
